Extracted grading from teacher-angular-js-module.
Moved component grading into component-grading-module.
Moved data export definitions into new data-export-module.
Started moving state definitions into own module files.

#150
